% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_images_to_hdfs}
\alias{copy_images_to_hdfs}
\title{Copy images required for the image classification demo to HDFS.}
\usage{
copy_images_to_hdfs(dest_dir = NULL)
}
\arguments{
\item{dest_dir}{The destination directory on HDFS containing images required
for the demo (e.g., "/dogs-vs-cats"). If unspecified, then it will be a
randomly generated path of the following form: /dogs-vs-cats_<UUID>.}
}
\value{
The destination directory on HDFS (e.g., "/dogs-vs-cats_<UUID>").
}
\description{
Copy train- and test- images required for the dogs-vs-cats sparklyr image
classification demo to HDFS.
}
\details{
The directory structure will be the following:

<dest-dir>/
  \
   train/
   |\
   | dogs/
   |  \
   |   dog.*.jpg
   |
   |\
   | cats/
   |  \
   |   cat.*.jpg
   |
   test/
   |\
   | dogs/
   |  \
   |   dog.*.jpg
    \
     cats/
      \
       cat.*.jpg
}
