% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{run_demo}
\alias{run_demo}
\title{Run the image classification demo.}
\usage{
run_demo(sc)
}
\arguments{
\item{sc}{The Spark connection.}
}
\value{
The trained image classification model.
}
\description{
Run the dogs-vs-cats sparklyr image classification demo.
}
